lang_portal_backend/requirements.txt
Django==5.0.2
djangorestframework==3.14.0
youtube-transcript-api==0.6.2
boto3==1.34.44
botocore==1.34.44
numpy==1.26.4
scikit-learn==1.4.1.post1
django-filter==23.5
sqlparse==0.4.4
asgiref==3.7.2
pytz==2024.1
django-cors-headers==4.3.1
python-dotenv==1.0.1
# LangChain and Vector Store
langchain>=0.1.9,<0.2.0
langchain-core>=0.1.53,<0.2.0
langchain-community>=0.0.24,<0.1.0
chromadb==0.4.22
hnswlib==0.7.0
pydantic<2.0
typing-extensions==4.9.0
protobuf==4.25.2
# Web and Utilities
streamlit==1.31.1
requests==2.31.0
tqdm==4.66.2
gTTS==2.5.1
PyYAML==6.0.1
tenacity==8.2.3